+++ lib/install_go.sh
@@ -1,16 +1,14 @@
 #!/bin/bash
 
-# Borrowing this script from the Node.js Buildpack.  
+# Borrowing this script from the Node.js Buildpack.
 # https://raw.githubusercontent.com/cloudfoundry/nodejs-buildpack/master/scripts/install_go.sh
 
 set -euo pipefail
 
-GO_VERSION="1.12.4"
+GO_VERSION="1.13.6"
 
-if [ $CF_STACK == "cflinuxfs2" ]; then
-    GO_SHA256="1a6d80b16a845f6a9692857a5978c0d69a89b58af4d50f66209435beafb07b5b"
-elif [ $CF_STACK == "cflinuxfs3" ]; then
-    GO_SHA256="e68279204493307782c51105c3dd5254ab066d0b5d9aafa3ce3a2878ebbef53f"
+if [ $CF_STACK == "cflinuxfs3" ]; then
+    GO_SHA256="56df080cd10ad7d827abb4e826cae46b66d68451f38bd1286610639d1601ead5"
 else
   echo "       **ERROR** Unsupported stack"
   echo "                 See https://docs.cloudfoundry.org/devguide/deploy-apps/stacks.html for more info"
@@ -21,7 +19,7 @@ export GoInstallDir="/tmp/go$GO_VERSION"
 mkdir -p $GoInstallDir
 
 if [ ! -f $GoInstallDir/go/bin/go ]; then
-  URL=https://buildpacks.cloudfoundry.org/dependencies/go/go${GO_VERSION}.linux-amd64-${CF_STACK}-${GO_SHA256:0:8}.tar.gz
+  URL=https://buildpacks.cloudfoundry.org/dependencies/go/go${GO_VERSION}.linux-amd64-${CF_STACK}-${GO_SHA256:0:8}.tgz
 
   echo "-----> Download go ${GO_VERSION}"
   curl -s -L --retry 15 --retry-delay 2 $URL -o /tmp/go.tar.gz
@@ -36,7 +34,7 @@ if [ ! -f $GoInstallDir/go/bin/go ]; then
   tar xzf /tmp/go.tar.gz -C $GoInstallDir
   rm /tmp/go.tar.gz
 fi
-if [ ! -f $GoInstallDir/go/bin/go ]; then
+if [ ! -f $GoInstallDir/bin/go ]; then
   echo "       **ERROR** Could not download go"
   exit 1
 fi